make functions double
git-svn-id: http://emulsion.na.infn.it/svn/fedra/trunk@207 96872706-0045-0410-9534-84be8447c86c

+++ src/libEmath/EdbMath.h
@@ -18,8 +18,8 @@ class EdbMath {
   EdbMath(){}  
   virtual ~EdbMath(){}
 
-  static float Magnitude3( float Point1[3], float Point2[3] );
-  static float DistancePointLine3( float Point[3], float LineStart[3], float LineEnd[3], bool inside );
+  static double Magnitude3( float Point1[3], float Point2[3] );
+  static double DistancePointLine3( float Point[3], float LineStart[3], float LineEnd[3], bool inside );
 
   ClassDef(EdbMath,1)  // general matematical algorithms
 };
